requirements.txt
africastalking==1.2.7
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contextlib2==21.6.0
cryptography==42.0.2
Django==5.0.2
django-cors-headers==4.2.0
django-nose==1.4.7
django-oauth-toolkit==2.3.0
django-tenants==3.6.1
djangorestframework==3.14.0
idna==3.6
isort==5.13.2
jwcrypto==1.5.3
kombu==5.3.5
Markdown==3.5.2
nose==1.3.7
oauthlib==3.2.2
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
schema==0.7.5
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
